% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dflow.R
\name{dflow_add_package}
\alias{dflow_add_package}
\title{dflow_add_package}
\usage{
dflow_add_package(package)
}
\arguments{
\item{package}{Name of the desired package to be added, as a string.}
}
\value{
updates a file and loads a package.
}
\description{
Adds a package to the ./packages.R file
}
\details{
Adds a desired package to the ./packages.R file, and loads the package into the interactive environment. Does not take a snapshot of the renv.
}
\author{
Sharleen Weatherley
}
