% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loadd_functions}
\alias{loadd_functions}
\title{Loads all current packages and sources all functions}
\usage{
loadd_functions()
}
\value{
All scripts and functions
}
\description{
Loads all current packages and sources all functions
}
