% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dflow.R
\name{use_dflow_readme}
\alias{use_dflow_readme}
\title{use_dflow_readme}
\usage{
use_dflow_readme()
}
\value{
nothing, creates a file.
}
\description{
Use a starter README
}
\details{
Drop a starter README in the current working directory.
}
\author{
Sharleen Weatherley
}
