% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dflow.R
\name{use_dflow_rprofile}
\alias{use_dflow_rprofile}
\title{use_dflow_rprofile}
\usage{
use_dflow_rprofile()
}
\value{
nothing, creates a file.
}
\description{
Use a starter project .Rprofile
}
\details{
Drop a starter .Rprofile in the current working directory,which sources the home .Rprofile, but allows for project-specific .Rpfoile settings.
}
\author{
Miles McBain
}
